refactor: Create Request.hpp #22
- 클래스에 준하는 구조체이기 때문에 별도의 hpp 파일로 분리
