+++ include/Request.hpp
@@ -1,0 +1,63 @@
+#ifndef REQUEST_HPP
+#define REQUEST_HPP
+
+#include <vector>
+#include <map>
+#include <string>
+
+enum StatusCode
+{
+  NO_PROBLEM = 0,
+  BAD_REQUEST_400 = 400,
+};
+
+enum ValidationStatus
+{
+  READY,
+  ON_HEADER,
+  ON_BODY,
+  COMPLETE
+};
+
+struct Request
+{
+  ValidationStatus validation_phase;  // 파싱 단계를 알려주는 변수
+  StatusCode status;  // 오류 발생 여부를 알려주는 변수
+  std::string method;
+  std::string uri;
+  std::string http_version;
+  std::map<std::string, std::string> headers;
+  std::vector<char> body;
+
+  // Default Constructor
+  Request(void) : validation_phase(READY), status(NO_PROBLEM) {}
+
+  // Destructor
+  ~Request(void) {}
+
+  // Copy Constructor
+  Request(const Request& src)
+  {
+    if (this != &src)
+    {
+      *this = src;
+    }
+  }
+
+  // Copy Assignment
+  Request& operator=(const Request& rhs)
+  {
+    if (this != &rhs)
+    {
+      validation_phase = rhs.validation_phase;
+      uri = rhs.uri;
+      http_version = rhs.http_version;
+      headers = rhs.headers;
+      body = rhs.body;
+      status = rhs.status;
+    }
+    return (*this);
+  }
+};
+
+#endif
